Refactor credit scoring and advice functionality
- Moved credit-related functions from `auth` module to a new `credit_info` module for better organization.
- Updated test cases in `test_credit_app.py` to reflect the changes in function imports and ensure proper testing of credit scoring and advice functionalities.
- Adjusted function calls in tests to use the new `credit_info` module.
